Add files via upload
Faculty of Engineering and Technology Faculty of Computer Science and Engineering Department of First Year

Synopsis on “” for Project

Topic of the Project: Creating a PowerPoint presentation for your college project can be a powerful way to showcase your research and ideas. This guide will provide you with essential tips and strategies for crafting compelling and effective presentations.

Objective and Scope: The Sneaking website is for the students or intern who don’t
know how to make presentation or what are the point in it?

Process Description: Explaining the project- both the product and process to the Evaluators. The presentation complements the project documentation and the project demo it gives Evaluators a channel to clear up doubts by asking question on the spot.

Resources and Limitations: For the formation of this website we are using HTML,CSS,PYTHON (DSA) etc. language as it is very convenient language. For now this website is only browser based but in future we will develop it in an app based system.

Conclusion: The conclusion is that students and freshers should get all the information on our website whether it is presentation or resume, they do not need to go anywhere, they can find all the needful on our website.

Name of Student: BHAVYA ALAG
Year \ Semester: 1st YEAR (1st Semester) Registration No.:2024PUFCEBCEX19677 Details of Team Members: 1.DEEPALI Tanwar (1st YEAR (1st Semester))2024/17526 2. Bhoomi Arora (1st YEAR (1st Semester))2024/16914 3. Dushyant Tinker (1st YEAR (1st Semester))2024/17170